requirements.txt
asgiref==3.5.2
Django==4.1.1
django-nested-admin==4.0.0
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
